+++ src/library/audio/openal/al.cpp
@@ -37,7 +37,7 @@ void alSetError(ALenum error)
 
 ALenum alGetError(ALvoid)
 {
-    debuglog(LCF_OPENAL, __func__, " call, returning ", alError);
+    debuglog(LCF_SOUND, __func__, " call, returning ", alError);
     ALenum err = alError;
     alError = AL_NO_ERROR;
     return err;
@@ -45,7 +45,7 @@ ALenum alGetError(ALvoid)
 
 ALboolean alIsExtensionPresent(const ALchar *extname)
 {
-    debuglog(LCF_OPENAL, __func__, " call with extname ", extname);
+    debuglog(LCF_SOUND, __func__, " call with extname ", extname);
 
     if (strcmp(extname, ALC_EXT_EFX_NAME) == 0) {
         return AL_TRUE;
@@ -62,7 +62,7 @@ ALboolean alIsExtensionPresent(const ALchar *extname)
 
 void* alGetProcAddress(const ALchar *fname)
 {
-    debuglog(LCF_OPENAL, __func__, " call with name ", fname);
+    debuglog(LCF_SOUND, __func__, " call with name ", fname);
 
     CHECK_RETURN_MY_FUNCTION(alGenEffects)
     CHECK_RETURN_MY_FUNCTION(alDeleteEffects)
@@ -105,7 +105,7 @@ void* alGetProcAddress(const ALchar *fname)
 
 void alGenBuffers(ALsizei n, ALuint *buffers)
 {
-    debuglog(LCF_OPENAL, __func__, " call - generate ", n, " buffers");
+    debuglog(LCF_SOUND, __func__, " call - generate ", n, " buffers");
 
     std::lock_guard<std::mutex> lock(audiocontext.mutex);
     for (int i=0; i<n; i++) {
@@ -118,7 +118,7 @@ void alGenBuffers(ALsizei n, ALuint *buffers)
 
 void alDeleteBuffers(ALsizei n, ALuint *buffers)
 {
-    debuglog(LCF_OPENAL, __func__, " call - delete ", n, " buffers");
+    debuglog(LCF_SOUND, __func__, " call - delete ", n, " buffers");
 
     std::lock_guard<std::mutex> lock(audiocontext.mutex);
     for (int i=0; i<n; i++) {
@@ -135,14 +135,14 @@ void alDeleteBuffers(ALsizei n, ALuint *buffers)
 
 ALboolean alIsBuffer(ALuint buffer)
 {
-    DEBUGLOGCALL(LCF_OPENAL);
+    DEBUGLOGCALL(LCF_SOUND);
     std::lock_guard<std::mutex> lock(audiocontext.mutex);
     return audiocontext.isBuffer(buffer);
 }
 
 void alBufferData(ALuint buffer, ALenum format, const ALvoid *data, ALsizei size, ALsizei freq)
 {
-    debuglog(LCF_OPENAL, __func__, " call - copy buffer data of format ", format, ", size ", size, " and frequency ", freq, " into buffer ", buffer);
+    debuglog(LCF_SOUND, __func__, " call - copy buffer data of format ", format, ", size ", size, " and frequency ", freq, " into buffer ", buffer);
 
     std::lock_guard<std::mutex> lock(audiocontext.mutex);
 
@@ -201,7 +201,7 @@ void alBufferData(ALuint buffer, ALenum format, const ALvoid *data, ALsizei size
                 ab->blockSamples = 64;
             break;
         default:
-            debuglog(LCF_OPENAL | LCF_ERROR, "Unsupported format: ", format);
+            debuglog(LCF_SOUND | LCF_ERROR, "Unsupported format: ", format);
             return;
     }
 
@@ -221,25 +221,25 @@ void alBufferData(ALuint buffer, ALenum format, const ALvoid *data, ALsizei size
 
 void alBufferf(ALuint buffer, ALenum param, ALfloat value)
 {
-    DEBUGLOGCALL(LCF_OPENAL);
-    debuglog(LCF_OPENAL, "Operation not supported");
+    DEBUGLOGCALL(LCF_SOUND);
+    debuglog(LCF_SOUND, "Operation not supported");
 }
 
 void alBuffer3f(ALuint buffer, ALenum param, ALfloat value1, ALfloat value2, ALfloat value3)
 {
-    DEBUGLOGCALL(LCF_OPENAL);
-    debuglog(LCF_OPENAL, "Operation not supported");
+    DEBUGLOGCALL(LCF_SOUND);
+    debuglog(LCF_SOUND, "Operation not supported");
 }
 
 void alBufferfv(ALuint buffer, ALenum param, const ALfloat *values)
 {
-    DEBUGLOGCALL(LCF_OPENAL);
-    debuglog(LCF_OPENAL, "Operation not supported");
+    DEBUGLOGCALL(LCF_SOUND);
+    debuglog(LCF_SOUND, "Operation not supported");
 }
 
 void alBufferi(ALuint buffer, ALenum param, ALint value)
 {
-    DEBUGLOGCALL(LCF_OPENAL);
+    DEBUGLOGCALL(LCF_SOUND);
     std::lock_guard<std::mutex> lock(audiocontext.mutex);
     auto ab = audiocontext.getBuffer(buffer);
     if (ab == nullptr) {
@@ -249,25 +249,25 @@ void alBufferi(ALuint buffer, ALenum param, ALint value)
 
     switch(param) {
         case AL_UNPACK_BLOCK_ALIGNMENT_SOFT:
-            debuglog(LCF_OPENAL, "  Set block alignment ", value);
+            debuglog(LCF_SOUND, "  Set block alignment ", value);
             ab->blockSamples = value;
             ab->update();
             break;
         default:
-            debuglog(LCF_OPENAL, "  Operation not supported: ", param);
+            debuglog(LCF_SOUND, "  Operation not supported: ", param);
             return;
     }
 }
 
 void alBuffer3i(ALuint buffer, ALenum param, ALint value1, ALint value2, ALint value3)
 {
-    DEBUGLOGCALL(LCF_OPENAL);
-    debuglog(LCF_OPENAL, "Operation not supported");
+    DEBUGLOGCALL(LCF_SOUND);
+    debuglog(LCF_SOUND, "Operation not supported");
 }
 
 void alBufferiv(ALuint buffer, ALenum param, const ALint *values)
 {
-    DEBUGLOGCALL(LCF_OPENAL);
+    DEBUGLOGCALL(LCF_SOUND);
     if (values == nullptr) {
         alSetError(AL_INVALID_VALUE);
         return;
@@ -285,12 +285,12 @@ void alBufferiv(ALuint buffer, ALenum param, const ALint *values)
             break;
         case AL_LOOP_POINTS_SOFT:
             /* TODO: Generate the errors */
-            debuglog(LCF_OPENAL, "  Set loop points ", values[0], " -> ", values[1]);
+            debuglog(LCF_SOUND, "  Set loop points ", values[0], " -> ", values[1]);
             ab->loop_point_beg = values[0];
             ab->loop_point_end = values[1];
             break;
         default:
-            debuglog(LCF_OPENAL, "  Operation not supported: ", param);
+            debuglog(LCF_SOUND, "  Operation not supported: ", param);
             return;
     }
 }
@@ -298,7 +298,7 @@ void alBufferiv(ALuint buffer, ALenum param, const ALint *values)
 
 void alGetBufferi(ALuint buffer, ALenum pname, ALint *value)
 {
-    DEBUGLOGCALL(LCF_OPENAL);
+    DEBUGLOGCALL(LCF_SOUND);
 
     if (value == nullptr) {
         return;
@@ -314,23 +314,23 @@ void alGetBufferi(ALuint buffer, ALenum pname, ALint *value)
     switch(pname) {
         case AL_FREQUENCY:
             *value = ab->frequency;
-            debuglog(LCF_OPENAL, "  Get frequency of ", *value);
+            debuglog(LCF_SOUND, "  Get frequency of ", *value);
             return;
         case AL_BITS:
             *value = ab->bitDepth;
-            debuglog(LCF_OPENAL, "  Get bit depth of ", *value);
+            debuglog(LCF_SOUND, "  Get bit depth of ", *value);
             return;
         case AL_CHANNELS:
             *value = ab->nbChannels;
-            debuglog(LCF_OPENAL, "  Get channel number of ", *value);
+            debuglog(LCF_SOUND, "  Get channel number of ", *value);
             return;
         case AL_SIZE:
             *value = ab->size;
-            debuglog(LCF_OPENAL, "  Get size of ", *value);
+            debuglog(LCF_SOUND, "  Get size of ", *value);
             return;
         case AL_UNPACK_BLOCK_ALIGNMENT_SOFT:
             *value = ab->blockSamples;
-            debuglog(LCF_OPENAL, "  Get block alignment of ", *value);
+            debuglog(LCF_SOUND, "  Get block alignment of ", *value);
             return;
         default:
             alSetError(AL_INVALID_VALUE);
@@ -340,7 +340,7 @@ void alGetBufferi(ALuint buffer, ALenum pname, ALint *value)
 
 void alGetBufferiv(ALuint buffer, ALenum pname, ALint *values)
 {
-    DEBUGLOGCALL(LCF_OPENAL);
+    DEBUGLOGCALL(LCF_SOUND);
 
     if (values == nullptr) {
         return;
@@ -356,24 +356,24 @@ void alGetBufferiv(ALuint buffer, ALenum pname, ALint *values)
     switch(pname) {
         case AL_FREQUENCY:
             *values = ab->frequency;
-            debuglog(LCF_OPENAL, "  Get frequency of ", *values);
+            debuglog(LCF_SOUND, "  Get frequency of ", *values);
             return;
         case AL_BITS:
             *values = ab->bitDepth;
-            debuglog(LCF_OPENAL, "  Get bit depth of ", *values);
+            debuglog(LCF_SOUND, "  Get bit depth of ", *values);
             return;
         case AL_CHANNELS:
             *values = ab->nbChannels;
-            debuglog(LCF_OPENAL, "  Get channel number of ", *values);
+            debuglog(LCF_SOUND, "  Get channel number of ", *values);
             return;
         case AL_SIZE:
             *values = ab->size;
-            debuglog(LCF_OPENAL, "  Get size of ", *values);
+            debuglog(LCF_SOUND, "  Get size of ", *values);
             return;
         case AL_LOOP_POINTS_SOFT:
             values[0] = ab->loop_point_beg;
             values[1] = ab->loop_point_end;
-            debuglog(LCF_OPENAL, "  Get loop points ", values[0], " -> ", values[1]);
+            debuglog(LCF_SOUND, "  Get loop points ", values[0], " -> ", values[1]);
             break;
         default:
             alSetError(AL_INVALID_VALUE);
@@ -385,7 +385,7 @@ void alGetBufferiv(ALuint buffer, ALenum pname, ALint *values)
 
 void alGenSources(ALsizei n, ALuint *sources)
 {
-    debuglog(LCF_OPENAL, __func__, " call - generate ", n, " sources");
+    debuglog(LCF_SOUND, __func__, " call - generate ", n, " sources");
     std::lock_guard<std::mutex> lock(audiocontext.mutex);
 	for (int i=0; i<n; i++) {
 		int id = audiocontext.createSource();
@@ -401,7 +401,7 @@ void alGenSources(ALsizei n, ALuint *sources)
 
 void alDeleteSources(ALsizei n, ALuint *sources)
 {
-    debuglog(LCF_OPENAL, __func__, " call - delete ", n, " sources");
+    debuglog(LCF_SOUND, __func__, " call - delete ", n, " sources");
     std::lock_guard<std::mutex> lock(audiocontext.mutex);
 	for (int i=0; i<n; i++) {
         /* Check if all sources exist before removing any. */
@@ -421,14 +421,14 @@ void alDeleteSources(ALsizei n, ALuint *sources)
 
 ALboolean alIsSource(ALuint source)
 {
-    DEBUGLOGCALL(LCF_OPENAL);
+    DEBUGLOGCALL(LCF_SOUND);
     std::lock_guard<std::mutex> lock(audiocontext.mutex);
 	return audiocontext.isSource(source);
 }
 
 void alSourcef(ALuint source, ALenum param, ALfloat value)
 {
-    debuglogstdio(LCF_OPENAL, "%s called with source %d", __func__, source);
+    debuglogstdio(LCF_SOUND, "%s called with source %d", __func__, source);
     std::lock_guard<std::mutex> lock(audiocontext.mutex);
     auto as = audiocontext.getSource(source);
     if (!as) {
@@ -440,18 +440,18 @@ void alSourcef(ALuint source, ALenum param, ALfloat value)
     switch(param) {
         case AL_GAIN:
             as->volume = value;
-            debuglog(LCF_OPENAL, "  Set gain of ", value);
+            debuglog(LCF_SOUND, "  Set gain of ", value);
             break;
         case AL_PITCH:
             if (as->pitch != value) {
                 as->dirty();
             }
             as->pitch = value;
-            debuglog(LCF_OPENAL, "  Set pitch of ", value);
+            debuglog(LCF_SOUND, "  Set pitch of ", value);
             break;
         case AL_REFERENCE_DISTANCE:
             if (value != 1.0) {
-                debuglog(LCF_OPENAL, "  Set reference distance to ", value, ". Operation not supported");
+                debuglog(LCF_SOUND, "  Set reference distance to ", value, ". Operation not supported");
             }
             break;
         case AL_MIN_GAIN:
@@ -462,7 +462,7 @@ void alSourcef(ALuint source, ALenum param, ALfloat value)
         case AL_CONE_INNER_ANGLE:
         case AL_CONE_OUTER_ANGLE:
         case AL_AUXILIARY_SEND_FILTER:
-            debuglog(LCF_OPENAL, "Operation not supported: ", param);
+            debuglog(LCF_SOUND, "Operation not supported: ", param);
             break;
         case AL_SEC_OFFSET:
             /* We fetch the buffer format of the source.
@@ -470,7 +470,7 @@ void alSourcef(ALuint source, ALenum param, ALfloat value)
              */
             if (! as->buffer_queue.empty()) {
                 ab = as->buffer_queue[0];
-                debuglog(LCF_OPENAL, "  Set position of ", value, " seconds");
+                debuglog(LCF_SOUND, "  Set position of ", value, " seconds");
                 value *= (ALfloat) ab->frequency;
                 as->setPosition((int)value);
             }
@@ -479,19 +479,19 @@ void alSourcef(ALuint source, ALenum param, ALfloat value)
             /* We fetch the buffer format of the source.
              * Normally, all buffers from a queue share the exact same format.
              */
-            debuglog(LCF_OPENAL, "  Set position of ", value, " samples");
+            debuglog(LCF_SOUND, "  Set position of ", value, " samples");
             as->setPosition((int)value);
             break;
         case AL_BYTE_OFFSET:
             if (! as->buffer_queue.empty()) {
                 ab = as->buffer_queue[0];
                 value /= (ALfloat) ab->alignSize;
-                debuglog(LCF_OPENAL, "  Set position of ", value, " bytes");
+                debuglog(LCF_SOUND, "  Set position of ", value, " bytes");
                 as->setPosition((int)value);
             }
             break;
         default:
-            debuglog(LCF_OPENAL, "  Unknown param ", param);
+            debuglog(LCF_SOUND, "  Unknown param ", param);
             alSetError(AL_INVALID_OPERATION);
             return;
     }
@@ -499,23 +499,23 @@ void alSourcef(ALuint source, ALenum param, ALfloat value)
 
 void alSource3f(ALuint source, ALenum param, ALfloat v1, ALfloat v2, ALfloat v3)
 {
-    debuglogstdio(LCF_OPENAL, "%s called with source %d", __func__, source);
+    debuglogstdio(LCF_SOUND, "%s called with source %d", __func__, source);
     switch(param) {
         case AL_DIRECTION:
-            debuglogstdio(LCF_OPENAL, "Setting direction not supported");
+            debuglogstdio(LCF_SOUND, "Setting direction not supported");
             break;
         case AL_VELOCITY:
-            debuglogstdio(LCF_OPENAL, "Setting velocity not supported");
+            debuglogstdio(LCF_SOUND, "Setting velocity not supported");
             break;
         default:
-            debuglog(LCF_OPENAL, "Operation not supported");
+            debuglog(LCF_SOUND, "Operation not supported");
             break;
     }
 }
 
 void alSourcefv(ALuint source, ALenum param, ALfloat *values)
 {
-    debuglogstdio(LCF_OPENAL, "%s called with source %d", __func__, source);
+    debuglogstdio(LCF_SOUND, "%s called with source %d", __func__, source);
     if (values == nullptr) {
         alSetError(AL_INVALID_VALUE);
         return;
@@ -523,7 +523,7 @@ void alSourcefv(ALuint source, ALenum param, ALfloat *values)
 
     switch(param) {
         case AL_POSITION:
-            debuglogstdio(LCF_OPENAL, "Setting position not supported");
+            debuglogstdio(LCF_SOUND, "Setting position not supported");
             break;
         default:
             alSourcef(source, param, *values);
@@ -533,7 +533,7 @@ void alSourcefv(ALuint source, ALenum param, ALfloat *values)
 
 void alSourcei(ALuint source, ALenum param, ALint value)
 {
-    debuglogstdio(LCF_OPENAL, "%s called with source %d", __func__, source);
+    debuglogstdio(LCF_SOUND, "%s called with source %d", __func__, source);
     std::lock_guard<std::mutex> lock(audiocontext.mutex);
     auto as = audiocontext.getSource(source);
     if (!as) {
@@ -544,7 +544,7 @@ void alSourcei(ALuint source, ALenum param, ALint value)
     std::shared_ptr<AudioBuffer> ab;
     switch(param) {
         case AL_LOOPING:
-            debuglog(LCF_OPENAL, "  Set looping of ", value);
+            debuglog(LCF_SOUND, "  Set looping of ", value);
             if (value == AL_TRUE)
                 as->looping = true;
             else if (value == AL_FALSE)
@@ -563,7 +563,7 @@ void alSourcei(ALuint source, ALenum param, ALint value)
             if (value == 0) {
                 /* Unbind buffer from source */
                 as->init();
-                debuglog(LCF_OPENAL, "  Unbind buffer");
+                debuglog(LCF_SOUND, "  Unbind buffer");
             }
             else {
                 ab = audiocontext.getBuffer(value);
@@ -574,7 +574,7 @@ void alSourcei(ALuint source, ALenum param, ALint value)
                 as->init();
                 as->buffer_queue.push_back(ab);
                 as->source = AudioSource::SOURCE_STATIC;
-                debuglog(LCF_OPENAL, "  Bind to buffer ", value);
+                debuglog(LCF_SOUND, "  Bind to buffer ", value);
             }
             break;
         case AL_SOURCE_RELATIVE:
@@ -584,7 +584,7 @@ void alSourcei(ALuint source, ALenum param, ALint value)
         case AL_DIRECT_FILTER_GAINHF_AUTO:
         case AL_AUXILIARY_SEND_FILTER_GAIN_AUTO:
         case AL_AUXILIARY_SEND_FILTER_GAINHF_AUTO:
-            debuglog(LCF_OPENAL, "Operation not supported: ", param);
+            debuglog(LCF_SOUND, "Operation not supported: ", param);
             break;
         case AL_SEC_OFFSET:
             /* We fetch the buffer format of the source.
@@ -592,7 +592,7 @@ void alSourcei(ALuint source, ALenum param, ALint value)
              */
             if (! as->buffer_queue.empty()) {
                 ab = as->buffer_queue[0];
-                debuglog(LCF_OPENAL, "  Set position of ", value, " seconds");
+                debuglog(LCF_SOUND, "  Set position of ", value, " seconds");
                 value *= static_cast<ALint>(ab->frequency);
                 as->setPosition(static_cast<int>(value));
             }
@@ -601,19 +601,19 @@ void alSourcei(ALuint source, ALenum param, ALint value)
             /* We fetch the buffer format of the source.
              * Normally, all buffers from a queue share the exact same format.
              */
-            debuglog(LCF_OPENAL, "  Set position of ", value, " samples");
+            debuglog(LCF_SOUND, "  Set position of ", value, " samples");
             as->setPosition(static_cast<int>(value));
             break;
         case AL_BYTE_OFFSET:
             if (! as->buffer_queue.empty()) {
                 ab = as->buffer_queue[0];
                 value /= static_cast<ALint>(ab->alignSize);
-                debuglog(LCF_OPENAL, "  Set position of ", value, " bytes");
+                debuglog(LCF_SOUND, "  Set position of ", value, " bytes");
                 as->setPosition(static_cast<int>(value));
             }
             break;
         default:
-            debuglog(LCF_OPENAL, "  Unknown param ", param);
+            debuglog(LCF_SOUND, "  Unknown param ", param);
             alSetError(AL_INVALID_OPERATION);
             return;
     }
@@ -621,18 +621,18 @@ void alSourcei(ALuint source, ALenum param, ALint value)
 
 void alSource3i(ALuint source, ALenum param, ALint v1, ALint v2, ALint v3)
 {
-    debuglogstdio(LCF_OPENAL, "%s called with source %d", __func__, source);
+    debuglogstdio(LCF_SOUND, "%s called with source %d", __func__, source);
     switch(param) {
         case AL_AUXILIARY_SEND_FILTER:
-            debuglog(LCF_OPENAL, "Operation not supported");
+            debuglog(LCF_SOUND, "Operation not supported");
             return;
     }
-    debuglog(LCF_OPENAL, "Operation not supported");
+    debuglog(LCF_SOUND, "Operation not supported");
 }
 
 void alSourceiv(ALuint source, ALenum param, ALint *values)
 {
-    debuglogstdio(LCF_OPENAL, "%s called with source %d", __func__, source);
+    debuglogstdio(LCF_SOUND, "%s called with source %d", __func__, source);
     if (values == nullptr) {
         alSetError(AL_INVALID_VALUE);
         return;
@@ -642,7 +642,7 @@ void alSourceiv(ALuint source, ALenum param, ALint *values)
 
 void alGetSourcef(ALuint source, ALenum param, ALfloat *value)
 {
-    DEBUGLOGCALL(LCF_OPENAL);
+    DEBUGLOGCALL(LCF_SOUND);
 
     if (value == nullptr) {
         return;
@@ -657,11 +657,11 @@ void alGetSourcef(ALuint source, ALenum param, ALfloat *value)
     switch(param) {
         case AL_GAIN:
             *value = as->volume;
-            debuglog(LCF_OPENAL, "  Get gain of ", *value);
+            debuglog(LCF_SOUND, "  Get gain of ", *value);
             break;
         case AL_PITCH:
             *value = as->pitch;
-            debuglog(LCF_OPENAL, "  Get pitch of ", *value);
+            debuglog(LCF_SOUND, "  Get pitch of ", *value);
             break;
         case AL_MIN_GAIN:
         case AL_MAX_GAIN:
@@ -674,7 +674,7 @@ void alGetSourcef(ALuint source, ALenum param, ALfloat *value)
         case AL_AIR_ABSORPTION_FACTOR:
         case AL_ROOM_ROLLOFF_FACTOR:
         case AL_CONE_OUTER_GAINHF:
-            debuglog(LCF_OPENAL, "Operation not supported: ", param);
+            debuglog(LCF_SOUND, "Operation not supported: ", param);
             break;
         case AL_SEC_OFFSET:
             /* We fetch the buffer format of the source.
@@ -683,12 +683,12 @@ void alGetSourcef(ALuint source, ALenum param, ALfloat *value)
             if (! as->buffer_queue.empty()) {
                 ab = as->buffer_queue[0];
                 *value = static_cast<ALfloat>(as->getPosition()) / ab->frequency;
-                debuglog(LCF_OPENAL, "  Get position of ", *value, " seconds");
+                debuglog(LCF_SOUND, "  Get position of ", *value, " seconds");
             }
             break;
         case AL_SAMPLE_OFFSET:
             *value = (ALfloat) as->getPosition();
-            debuglog(LCF_OPENAL, "  Get position of ", *value, " samples");
+            debuglog(LCF_SOUND, "  Get position of ", *value, " samples");
             break;
         case AL_BYTE_OFFSET:
             /* We fetch the buffer format of the source.
@@ -697,11 +697,11 @@ void alGetSourcef(ALuint source, ALenum param, ALfloat *value)
             if (! as->buffer_queue.empty()) {
                 ab = as->buffer_queue[0];
                 *value = static_cast<ALfloat>(as->getPosition()) * ab->alignSize;
-                debuglog(LCF_OPENAL, "  Get position of ", *value, " bytes");
+                debuglog(LCF_SOUND, "  Get position of ", *value, " bytes");
             }
             break;
         default:
-            debuglog(LCF_OPENAL, "  Unknown param ", param);
+            debuglog(LCF_SOUND, "  Unknown param ", param);
             alSetError(AL_INVALID_OPERATION);
             return;
     }
@@ -709,19 +709,19 @@ void alGetSourcef(ALuint source, ALenum param, ALfloat *value)
 
 void alGetSource3f(ALuint source, ALenum param, ALfloat *v1, ALfloat *v2, ALfloat *v3)
 {
-    DEBUGLOGCALL(LCF_OPENAL);
-    debuglog(LCF_OPENAL, "Operation not supported");
+    DEBUGLOGCALL(LCF_SOUND);
+    debuglog(LCF_SOUND, "Operation not supported");
 }
 
 void alGetSourcefv(ALuint source, ALenum param, ALfloat *values)
 {
-    DEBUGLOGCALL(LCF_OPENAL);
+    DEBUGLOGCALL(LCF_SOUND);
     alGetSourcef(source, param, values);
 }
 
 void alGetSourcei(ALuint source, ALenum param, ALint *value)
 {
-    debuglog(LCF_OPENAL, __func__, " call for source ", source);
+    debuglog(LCF_SOUND, __func__, " call for source ", source);
 
     if (value == nullptr) {
         return;
@@ -747,20 +747,20 @@ void alGetSourcei(ALuint source, ALenum param, ALint *value)
             switch(as->state) {
                 case AudioSource::SOURCE_INITIAL:
                     *value = AL_INITIAL;
-                    debuglog(LCF_OPENAL, "  Get source state INITIAL");
+                    debuglog(LCF_SOUND, "  Get source state INITIAL");
                     break;
                 case AudioSource::SOURCE_PREPARED:
                 case AudioSource::SOURCE_PLAYING:
                     *value = AL_PLAYING;
-                    debuglog(LCF_OPENAL, "  Get source state PLAYING");
+                    debuglog(LCF_SOUND, "  Get source state PLAYING");
                     break;
                 case AudioSource::SOURCE_PAUSED:
                     *value = AL_PAUSED;
-                    debuglog(LCF_OPENAL, "  Get source state PAUSED");
+                    debuglog(LCF_SOUND, "  Get source state PAUSED");
                     break;
                 case AudioSource::SOURCE_STOPPED:
                     *value = AL_STOPPED;
-                    debuglog(LCF_OPENAL, "  Get source state STOPPED");
+                    debuglog(LCF_SOUND, "  Get source state STOPPED");
                     break;
             }
             break;
@@ -768,15 +768,15 @@ void alGetSourcei(ALuint source, ALenum param, ALint *value)
             switch(as->source) {
                 case AudioSource::SOURCE_UNDETERMINED:
                     *value = AL_UNDETERMINED;
-                    debuglog(LCF_OPENAL, "  Get source type UNDETERMINED");
+                    debuglog(LCF_SOUND, "  Get source type UNDETERMINED");
                     break;
                 case AudioSource::SOURCE_STATIC:
                     *value = AL_STATIC;
-                    debuglog(LCF_OPENAL, "  Get source type STATIC");
+                    debuglog(LCF_SOUND, "  Get source type STATIC");
                     break;
                 case AudioSource::SOURCE_STREAMING:
                     *value = AL_STREAMING;
-                    debuglog(LCF_OPENAL, "  Get source type STREAMING");
+                    debuglog(LCF_SOUND, "  Get source type STREAMING");
                     break;
                 default:
                     alSetError(AL_INVALID_VALUE);
@@ -787,11 +787,11 @@ void alGetSourcei(ALuint source, ALenum param, ALint *value)
         case AL_DIRECT_FILTER_GAINHF_AUTO:
         case AL_AUXILIARY_SEND_FILTER_GAIN_AUTO:
         case AL_AUXILIARY_SEND_FILTER_GAINHF_AUTO:
-            debuglog(LCF_OPENAL, "Operation not supported: ", param);
+            debuglog(LCF_SOUND, "Operation not supported: ", param);
             break;
         case AL_BUFFERS_QUEUED:
             *value = as->nbQueue();
-            debuglog(LCF_OPENAL, "  Get number of queued buffers of ", *value);
+            debuglog(LCF_SOUND, "  Get number of queued buffers of ", *value);
             break;
         case AL_BUFFERS_PROCESSED:
             if (as->state == AudioSource::SOURCE_STOPPED)
@@ -800,7 +800,7 @@ void alGetSourcei(ALuint source, ALenum param, ALint *value)
                 *value = 0;
             else
                 *value = as->nbQueueProcessed();
-            debuglog(LCF_OPENAL, "  Get number of processed queued buffers of ", *value);
+            debuglog(LCF_SOUND, "  Get number of processed queued buffers of ", *value);
             break;
         case AL_SEC_OFFSET:
             /* We fetch the buffer format of the source.
@@ -809,12 +809,12 @@ void alGetSourcei(ALuint source, ALenum param, ALint *value)
             if (! as->buffer_queue.empty()) {
                 ab = as->buffer_queue[0];
                 *value = as->getPosition() / ab->frequency;
-                debuglog(LCF_OPENAL, "  Get position of ", *value, " seconds");
+                debuglog(LCF_SOUND, "  Get position of ", *value, " seconds");
             }
             break;
         case AL_SAMPLE_OFFSET:
             *value = static_cast<ALint>(as->getPosition());
-            debuglog(LCF_OPENAL, "  Get position of ", *value, " samples");
+            debuglog(LCF_SOUND, "  Get position of ", *value, " samples");
             break;
         case AL_BYTE_OFFSET:
             /* We fetch the buffer format of the source.
@@ -823,11 +823,11 @@ void alGetSourcei(ALuint source, ALenum param, ALint *value)
             if (! as->buffer_queue.empty()) {
                 ab = as->buffer_queue[0];
                 *value = as->getPosition() * ab->alignSize;
-                debuglog(LCF_OPENAL, "  Get position of ", *value, " bytes");
+                debuglog(LCF_SOUND, "  Get position of ", *value, " bytes");
             }
             break;
         default:
-            debuglog(LCF_OPENAL, "  Unknown param ", param);
+            debuglog(LCF_SOUND, "  Unknown param ", param);
             alSetError(AL_INVALID_OPERATION);
             return;
     }
@@ -835,19 +835,19 @@ void alGetSourcei(ALuint source, ALenum param, ALint *value)
 
 void alGetSource3i(ALuint source, ALenum param, ALint *v1, ALint *v2, ALint *v3)
 {
-    DEBUGLOGCALL(LCF_OPENAL);
-    debuglog(LCF_OPENAL, "Operation not supported");
+    DEBUGLOGCALL(LCF_SOUND);
+    debuglog(LCF_SOUND, "Operation not supported");
 }
 
 void alGetSourceiv(ALuint source, ALenum param, ALint *values)
 {
-    DEBUGLOGCALL(LCF_OPENAL);
+    DEBUGLOGCALL(LCF_SOUND);
     alGetSourcei(source, param, values);
 }
 
 void alSourcePlay(ALuint source)
 {
-    DEBUGLOGCALL(LCF_OPENAL);
+    DEBUGLOGCALL(LCF_SOUND);
     std::lock_guard<std::mutex> lock(audiocontext.mutex);
     auto as = audiocontext.getSource(source);
     if (!as)
@@ -862,14 +862,14 @@ void alSourcePlay(ALuint source)
 
 void alSourcePlayv(ALsizei n, ALuint *sources)
 {
-    DEBUGLOGCALL(LCF_OPENAL);
+    DEBUGLOGCALL(LCF_SOUND);
     for (int i=0; i<n; i++)
         alSourcePlay(sources[i]);
 }
 
 void alSourcePause(ALuint source)
 {
-    DEBUGLOGCALL(LCF_OPENAL);
+    DEBUGLOGCALL(LCF_SOUND);
     std::lock_guard<std::mutex> lock(audiocontext.mutex);
     auto as = audiocontext.getSource(source);
     if (!as)
@@ -884,14 +884,14 @@ void alSourcePause(ALuint source)
 
 void alSourcePausev(ALsizei n, ALuint *sources)
 {
-    DEBUGLOGCALL(LCF_OPENAL);
+    DEBUGLOGCALL(LCF_SOUND);
     for (int i=0; i<n; i++)
         alSourcePause(sources[i]);
 }
 
 void alSourceStop(ALuint source)
 {
-    DEBUGLOGCALL(LCF_OPENAL);
+    DEBUGLOGCALL(LCF_SOUND);
     std::lock_guard<std::mutex> lock(audiocontext.mutex);
     auto as = audiocontext.getSource(source);
     if (!as)
@@ -907,14 +907,14 @@ void alSourceStop(ALuint source)
 
 void alSourceStopv(ALsizei n, ALuint *sources)
 {
-    DEBUGLOGCALL(LCF_OPENAL);
+    DEBUGLOGCALL(LCF_SOUND);
     for (int i=0; i<n; i++)
         alSourceStop(sources[i]);
 }
 
 void alSourceRewind(ALuint source)
 {
-    DEBUGLOGCALL(LCF_OPENAL);
+    DEBUGLOGCALL(LCF_SOUND);
     std::lock_guard<std::mutex> lock(audiocontext.mutex);
     auto as = audiocontext.getSource(source);
     if (!as)
@@ -930,14 +930,14 @@ void alSourceRewind(ALuint source)
 
 void alSourceRewindv(ALsizei n, ALuint *sources)
 {
-    DEBUGLOGCALL(LCF_OPENAL);
+    DEBUGLOGCALL(LCF_SOUND);
     for (int i=0; i<n; i++)
         alSourceRewind(sources[i]);
 }
 
 void alSourceQueueBuffers(ALuint source, ALsizei n, ALuint* buffers)
 {
-    debuglog(LCF_OPENAL, "Pushing ", n, " buffers in the queue of source ", source);
+    debuglog(LCF_SOUND, "Pushing ", n, " buffers in the queue of source ", source);
     std::lock_guard<std::mutex> lock(audiocontext.mutex);
     auto as = audiocontext.getSource(source);
     if (!as)
@@ -958,13 +958,13 @@ void alSourceQueueBuffers(ALuint source, ALsizei n, ALuint* buffers)
             return;
 
         as->buffer_queue.push_back(queue_ab);
-        debuglog(LCF_OPENAL, "  Pushed buffer ", buffers[i]);
+        debuglog(LCF_SOUND, "  Pushed buffer ", buffers[i]);
     }
 }
 
 void alSourceUnqueueBuffers(ALuint source, ALsizei n, ALuint* buffers)
 {
-    DEBUGLOGCALL(LCF_OPENAL);
+    DEBUGLOGCALL(LCF_SOUND);
     std::lock_guard<std::mutex> lock(audiocontext.mutex);
     auto as = audiocontext.getSource(source);
     if (!as)
@@ -981,7 +981,7 @@ void alSourceUnqueueBuffers(ALuint source, ALsizei n, ALuint* buffers)
         return;
     }
 
-    debuglog(LCF_OPENAL, "Unqueueing ", n, " buffers out of ", as->nbQueue());
+    debuglog(LCF_SOUND, "Unqueueing ", n, " buffers out of ", as->nbQueue());
 
     /* Save the id of the unqueued buffers */
     for (int i=0; i<n; i++) {
@@ -1000,57 +1000,57 @@ void alSourceUnqueueBuffers(ALuint source, ALsizei n, ALuint* buffers)
 
 void alListenerf(ALenum param, ALfloat value)
 {
-    DEBUGLOGCALL(LCF_OPENAL);
+    DEBUGLOGCALL(LCF_SOUND);
     if (param == AL_GAIN)
         audiocontext.outVolume = value;
 }
 
 void alListener3f(ALenum param, ALfloat v1, ALfloat v2, ALfloat v3)
 {
-    DEBUGLOGCALL(LCF_OPENAL);
+    DEBUGLOGCALL(LCF_SOUND);
     switch(param) {
     case AL_POSITION:
-        debuglog(LCF_OPENAL, "   Set Position to: ", v1, ", ", v2, ", ", v3);
+        debuglog(LCF_SOUND, "   Set Position to: ", v1, ", ", v2, ", ", v3);
         break;
     case AL_VELOCITY:
-        debuglog(LCF_OPENAL, "   Set Velocity to: ", v1, ", ", v2, ", ", v3);
+        debuglog(LCF_SOUND, "   Set Velocity to: ", v1, ", ", v2, ", ", v3);
         break;
     }
-    debuglog(LCF_OPENAL, "Operation not supported: ", param);
+    debuglog(LCF_SOUND, "Operation not supported: ", param);
 }
 
 void alListenerfv(ALenum param, ALfloat *values)
 {
-    DEBUGLOGCALL(LCF_OPENAL);
+    DEBUGLOGCALL(LCF_SOUND);
     switch(param) {
     case AL_ORIENTATION:
-        debuglog(LCF_OPENAL, "   Set Orientation to: ", values[0], ", ", values[1], ", ", values[2], ", ", values[3], ", ", values[4], ", ", values[5]);
+        debuglog(LCF_SOUND, "   Set Orientation to: ", values[0], ", ", values[1], ", ", values[2], ", ", values[3], ", ", values[4], ", ", values[5]);
         break;
     }
-    debuglog(LCF_OPENAL, "Operation not supported: ", param);
+    debuglog(LCF_SOUND, "Operation not supported: ", param);
 }
 
 void alListeneri(ALenum param, ALint value)
 {
-    DEBUGLOGCALL(LCF_OPENAL);
-    debuglog(LCF_OPENAL, "Operation not supported");
+    DEBUGLOGCALL(LCF_SOUND);
+    debuglog(LCF_SOUND, "Operation not supported");
 }
 
 void alListener3i(ALenum param, ALint v1, ALint v2, ALint v3)
 {
-    DEBUGLOGCALL(LCF_OPENAL);
-    debuglog(LCF_OPENAL, "Operation not supported");
+    DEBUGLOGCALL(LCF_SOUND);
+    debuglog(LCF_SOUND, "Operation not supported");
 }
 
 void alListeneriv(ALenum param, ALint *values)
 {
-    DEBUGLOGCALL(LCF_OPENAL);
-    debuglog(LCF_OPENAL, "Operation not supported");
+    DEBUGLOGCALL(LCF_SOUND);
+    debuglog(LCF_SOUND, "Operation not supported");
 }
 
 void alGetListenerf(ALenum param, ALfloat *value)
 {
-    DEBUGLOGCALL(LCF_OPENAL);
+    DEBUGLOGCALL(LCF_SOUND);
     if (param == AL_GAIN) {
         if (!value) {
             // alSetError(AL_INVALID_VALUE);
@@ -1062,35 +1062,35 @@ void alGetListenerf(ALenum param, ALfloat *value)
 
 void alGetListener3f(ALenum param, ALfloat *v1, ALfloat *v2, ALfloat *v3)
 {
-    DEBUGLOGCALL(LCF_OPENAL);
-    debuglog(LCF_OPENAL, "Operation not supported");
+    DEBUGLOGCALL(LCF_SOUND);
+    debuglog(LCF_SOUND, "Operation not supported");
 }
 
 void alGetListenerfv(ALenum param, ALfloat *values)
 {
-    DEBUGLOGCALL(LCF_OPENAL);
+    DEBUGLOGCALL(LCF_SOUND);
     if (param == AL_GAIN)
         alGetListenerf(param, values);
     else
-        debuglog(LCF_OPENAL, "Operation not supported");
+        debuglog(LCF_SOUND, "Operation not supported");
 }
 
 void alGetListeneri(ALenum param, ALint *value)
 {
-    DEBUGLOGCALL(LCF_OPENAL);
-    debuglog(LCF_OPENAL, "Operation not supported");
+    DEBUGLOGCALL(LCF_SOUND);
+    debuglog(LCF_SOUND, "Operation not supported");
 }
 
 void alGetListener3i(ALenum param, ALint *v1, ALint *v2, ALint *v3)
 {
-    DEBUGLOGCALL(LCF_OPENAL);
-    debuglog(LCF_OPENAL, "Operation not supported");
+    DEBUGLOGCALL(LCF_SOUND);
+    debuglog(LCF_SOUND, "Operation not supported");
 }
 
 void alGetListeneriv(ALenum param, ALint *values)
 {
-    DEBUGLOGCALL(LCF_OPENAL);
-    debuglog(LCF_OPENAL, "Operation not supported");
+    DEBUGLOGCALL(LCF_SOUND);
+    debuglog(LCF_SOUND, "Operation not supported");
 }
 
 }