+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 * Savestates can be compressed
 * Middle button to reset controller axis to 0
+* Hook getpid() used for prng seeding
 
 ### Changed
 ### Fixed